Merge heirarchy of kernel configs.
This takes the heirarchy of platform -> machine -> flavour to
allow a sensible final kernel configuration.  It will also allow
us to go through and simplify the kernel configs.

We merge the kernel config's such that platform gives us the baseline
and more specialized kconfig options get merged in over the top.